+++ src/django_typer/completers/__init__.py
@@ -17,16 +17,20 @@ types:
   :func:`~django_typer.completers.cmd.commands`.
 - **databases**: Complete Django database names using
   :func:`~django_typer.completers.db.databases`.
-- **import paths**: Complete Django database names using
+- **import paths**: Complete python import paths using
   :func:`~django_typer.completers.path.import_paths`.
-- **paths**: Complete Django database names using
+- **paths**: Complete file and directory paths using
   :func:`~django_typer.completers.path.paths`.
-- **directories**: Complete Django database names using
+- **directories**: Complete directories using
   :func:`~django_typer.completers.path.directories`.
 - **media paths**: Complete Django media paths using
   :func:`~django_typer.completers.path.media_paths`.
 - **static paths**: Complete Django static paths using
   :func:`~django_typer.completers.path.static_paths`.
+- **languages**: Complete Django language names using
+  :func:`~django_typer.completers.settings.languages`.
+- **settings**: Complete Django settings variable names using
+  :func:`~django_typer.completers.settings.setting`.
 """
 
 import typing as t
@@ -36,7 +40,15 @@ from click.core import ParameterSource
 from click.shell_completion import CompletionItem
 
 Completer = t.Callable[[Context, Parameter, str], t.List[CompletionItem]]
-Strings = t.Union[t.Sequence[str], t.KeysView[str], t.Generator[str, None, None]]
+
+ItemTuple = t.Union[t.Any, t.Tuple[t.Any, t.Any]]
+Strings = t.Union[
+    t.Sequence[ItemTuple],
+    t.KeysView[t.Any],
+    t.ValuesView[t.Any],
+    t.ItemsView[t.Any, t.Any],
+    t.Generator[ItemTuple, None, None],
+]
 
 
 def these_strings(
@@ -47,7 +59,8 @@ def these_strings(
     Get a completer that provides completion logic that matches the allowed strings.
 
     :param strings: A sequence of allowed strings or a callable that generates a
-        sequence of allowed strings.
+        sequence of allowed strings. If the sequence contains 2-tuples the second
+        element is used as the help text for the completion item.
     :param allow_duplicates: Whether or not to allow duplicate values. Defaults to
         False.
     :return: A completer function.
@@ -55,17 +68,21 @@ def these_strings(
 
     def complete(ctx: Context, param: Parameter, incomplete: str):
         present = []
+        items = []
         if (
             not allow_duplicates
             and param.name
             and ctx.get_parameter_source(param.name) is not ParameterSource.DEFAULT
         ):
             present = [value for value in (ctx.params.get(param.name) or [])]
-        return [
-            CompletionItem(item)
-            for item in (strings() if callable(strings) else strings)
-            if item.startswith(incomplete) and item not in present
-        ]
+
+        for item in strings() if callable(strings) else strings:
+            help = None
+            if isinstance(item, tuple):
+                item, help = str(item[0]), str(item[1])
+            if item.startswith(incomplete) and item not in present:
+                items.append(CompletionItem(item, help=help))
+        return items
 
     return complete