+++ tests/apps/test_app/models.py
@@ -1,4 +1,20 @@
 from django.db import models
+import django
+
+
+def get_ip_choices():
+    return [
+        ("192.168.0.1", "Private Residential"),
+        ("10.0.0.1", "Private Gateway"),
+        ("172.16.0.1", "Corporate Network Gateway"),
+        ("8.8.8.8", "Google DNS"),
+        ("127.0.0.1", "Localhost IPv4"),
+        ("::1", "Localhost IPv6"),
+        ("2001:4860:4860::8888", "Google DNS IPv6"),
+        ("fe80::1", "Link-Local IPv6"),
+        ("fd00::1", "Private Network IPv6"),
+        ("2001:db8::1", "Documentation Example IPv6"),
+    ]
 
 
 class ShellCompleteTester(models.Model):
@@ -31,3 +47,67 @@ class ShellCompleteTester(models.Model):
     time_field = models.TimeField(null=True, default=None, db_index=True)
 
     duration_field = models.DurationField(null=True, default=None, db_index=True)
+
+
+class ChoicesShellCompleteTester(models.Model):
+    CHAR_CHOICES = [
+        ("git", "Git"),
+        ("svn", "Subversion (SVN)"),
+        ("hg", "Mercurial (Hg)"),
+        ("bzr", "Bazaar (Bzr)"),
+        ("cvs", "CVS"),
+        ("perforce", "Perforce"),
+        ("fossil", "Fossil"),
+        ("darcs", "Darcs"),
+        ("monotone", "Monotone"),
+    ]
+
+    char_choice = models.CharField(
+        max_length=12,
+        choices=CHAR_CHOICES,
+        db_index=True,
+        default=None,
+        null=True,
+    )
+
+    INT_CHOICES = [
+        (1, "One"),
+        (2, "Two"),
+        (3, "Mercurial (Hg)"),
+        (4, "Bazaar (Bzr)"),
+        (5, "CVS"),
+        (6, "Perforce"),
+        (7, "Fossil"),
+        (8, "Darcs"),
+        (9, "Monotone"),
+        (10, "Ten"),
+        (11, "Eleven"),
+        (12, "Twelve"),
+    ]
+
+    int_choice = models.PositiveSmallIntegerField(
+        choices=INT_CHOICES,
+        db_index=True,
+        default=None,
+        null=True,
+    )
+
+    IP_CHOICES = [
+        ("192.168.0.1", "Private Residential"),
+        ("10.0.0.1", "Private Gateway"),
+        ("172.16.0.1", "Corporate Network Gateway"),
+        ("8.8.8.8", "Google DNS"),
+        ("127.0.0.1", "Localhost IPv4"),
+        ("::1", "Localhost IPv6"),
+        ("2001:4860:4860::8888", "Google DNS IPv6"),
+        ("fe80::1", "Link-Local IPv6"),
+        ("fd00::1", "Private Network IPv6"),
+        ("2001:db8::1", "Documentation Example IPv6"),
+    ]
+
+    ip_choice = models.GenericIPAddressField(
+        choices=IP_CHOICES,
+        db_index=True,
+        default=None,
+        null=True,
+    )